+++ js/utils.js
@@ -198,6 +198,11 @@ export async function downloadAvailabilityCalendar(webid, participants, solidFet
     const data = await getRDFasJson(url, frame, solidFetch);
     participants[webid].availabilityCalendar.data = data['@graph'] || data;
     participants[webid].availabilityCalendar.status = 'downloaded';
+
+    // This covers the case when no available slots are present in the response.
+    if (!Array.isArray(participants[webid].availabilityCalendar.data)) {
+      participants[webid].availabilityCalendar.data = [];
+    }
   } catch (e) {
     let error;